yc/js_ym/requirements.txt
pyyaml==5.3.1
jsonschema==4.4.0
pyrsistent==0.18.1